refactor: CarService 제거 및 역할 CarManager로 위임
CarManager 도메인으로 service의 역할을 대체할 수 있다 판단해 CarService를 제거하였음
