+++ src/main/java/racingcar/controller/RacingControllerFactory.java
@@ -2,17 +2,15 @@ package racingcar.controller;
 
 import racingcar.domain.CarManager;
 import racingcar.domain.Racing;
-import racingcar.utils.StringSplitter;
 import racingcar.view.OutputView;
 
 public class RacingControllerFactory {
 
     public static RacingController create() {
         CarManager carManager = new CarManager();
-        Racing racing = new Racing(carManager);
-        StringSplitter stringSplitter = new StringSplitter();
+        Racing racing = new Racing();
         OutputView outputView = new OutputView();
-        return new RacingController(racing, carManager, stringSplitter, outputView);
+        return new RacingController(racing, carManager, outputView);
     }
 
 }
